exe/requirements.txt
# Core Django
Django==5.2.7
python-decouple==3.8

# Database (using psycopg instead of psycopg2-binary for Python 3.13+ compatibility)
psycopg[binary]==3.2.3

# REST API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

# CORS handling
django-cors-headers==4.4.0

# Authentication
django-allauth==65.0.2

# AI Integration
google-generativeai==0.8.3
openai==1.54.3

# Async Tasks
celery==5.4.0
redis==5.2.0

# WebSocket Support
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Image Processing
Pillow==10.4.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Development
django-debug-toolbar==4.4.6
